refactor(cli): extract analyze_flow bounded-input prompts to _analyze_flow_prompts
Moves the three input-validation prompt helpers
(_ask_optional_float, _ask_optional_int, _ask_optional_choice) out of
amx/cli_support/commands/analyze_flow.py into a focused
amx/cli_support/_analyze_flow_prompts.py.

analyze_flow.py re-exports the three names so any internal call site
that imported the underscore form keeps working unchanged.

Tests in test_analyze_llm_overrides.py monkeypatched the 'ask' and
'ask_choice' names on the analyze_flow module — those patches move
to amx.cli_support._analyze_flow_prompts.ask / .ask_choice so they
actually reach the call sites.

analyze_flow.py: 2152 -> 2087 LOC (-65, -3%). First cut at the
analyze_flow split; the run-finalization + scope-mutation pieces
remain for follow-up extractions.
